+++ .envrc
@@ -1,0 +1,1 @@
+layout python_uv
